% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_comparis.R
\name{get_prices}
\alias{get_prices}
\title{Get prices for a given city and page}
\usage{
get_prices(city, page)
}
\arguments{
\item{city}{a length one character vector indicating a city.}

\item{page}{a length one numeric vector indicating a page number.}
}
\value{
a numeric vector of prices.
}
\description{
\code{get_pages()} scraps prices elements for a given Swiss city and page
from \url{https://www.comparis.ch} web-page and returns a numeric vector of
prices.
}
\examples{
prices <- get_pages("bern", 3)

}
\author{
Iegor Rudnytskyi \email{iegor.rudnytskyi@unil.ch}
}
