% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_circle}
\alias{make_circle}
\title{Plotting a circle}
\usage{
make_circle(center = c(0, 0), radius = 1, nb_step = 300,
  col = "darkblue", fill = NULL, lty = 1)
}
\arguments{
\item{center}{A \code{vector} of dimension 2 used to denote the center
of the circle (default: \code{center = c(0,0)}).}

\item{radius}{A \code{numeric} (positive) value used to indicate the radius of the
circle (default: \code{raduis = 1}).}

\item{nb_step}{A \code{numeric} value used the number of points used to
construct the circle (default: \code{nb_step = 300}).}

\item{col}{A \code{numeric} or \code{string} used to denote the color of
the border of the circle (default: \code{col = "darkblue"})}

\item{fill}{A \code{numeric} or \code{string} used to denote the number
the color for filling the circle. The default, \code{NULL}, is to leave
circle unfilled.}

\item{lty}{A \code{numeric} value used the denote the line type used for
the border of the circle (default: \code{lty = 1}). See \code{?par} for
more details.}
}
\description{
Plotting a circle
}
\examples{
# Example 1
plot(NA, xlim = c(-1.1,1.1), ylim = c(-1.1,1.1), xlab = "x", ylab = "y")
grid()
make_square(col = "blue")
make_circle(col = "red", lty = 2)

# Example 2
cols = hcl(h = seq(15, 375, length = 4), l = 65, c = 100)[1:3]
cols_trans = hcl(h = seq(15, 375, length = 4), l = 65, c = 100, alpha = 0.15)[1:3]
plot(NA, xlim = c(-310, 310), ylim = c(-310, 310), xlab = "x", ylab = "y")
grid()
make_circle(fill = "lightgrey", radius = 200, lty = 1)
make_circle(radius = 20, center = c(-300, 300), col = cols[1], fill = cols[1])
make_circle(radius = 20, center = c(300, 300), col =cols[2], fill = cols[2])
make_circle(radius = 20, center = c(0, -300), col = cols[3], fill = cols[3])
make_circle(radius = 449.2136, center = c(-300, 300),
            col = cols[1], fill = cols_trans[1], lty = 2)
make_circle(radius = 284.8427, center = c(300, 300),
            col = cols[2], fill = cols_trans[2], lty = 2)
make_circle(radius = 414.3106, center = c(0, -300),
            col = cols[3], fill = cols_trans[3], lty = 2)
}
\author{
Stephane Guerrier
}
