% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_fct.R
\name{make_files}
\alias{make_files}
\title{Internal function for RMarkdown app}
\usage{
make_files(input, rmd = FALSE)
}
\arguments{
\item{input}{A \code{data frame} coming fronm shiny app.}

\item{rmd}{A \code{boolean} to indicate if the output should be a rmd file or
simply print output.}
}
\description{
Create RMD file need for app or print content.
}
