% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_square}
\alias{make_square}
\title{Plotting a square}
\usage{
make_square(center = c(0, 0), side = 2, col = "darkblue", fill = NULL,
  lty = 1)
}
\arguments{
\item{center}{A \code{vector} of dimension 2 used to denote the center
of the square (default: \code{center = c(0,0)}).}

\item{side}{A \code{numeric} (positive) value used to indicate the length of one
side of the square (default: \code{side = 2}).}

\item{col}{A \code{numeric} or \code{string} used to denote the color of
the border of the square (default: \code{col = "darkblue"})}

\item{fill}{A \code{numeric} or \code{string} used to denote the number
the color for filling the square The default, \code{NULL}, is to leave
circle unfilled.}

\item{lty}{A \code{numeric} value used the denote the line type used for
the border of the square (default: \code{lty = 1}). See \code{?par} for
more details.}
}
\description{
Plotting a square
}
\examples{
cols = hcl(h = seq(15, 375, length = 4), l = 65, c = 100)[1:3]
cols_trans = hcl(h = seq(15, 375, length = 4), l = 65, c = 100, alpha = 0.15)[1:3]
plot(NA, xlim = c(-110, 110), ylim = c(-110, 110), xlab = "x", ylab = "y")
grid()
make_square(center = c(-50,50), side = 100, col = cols[1], lty = 4)
make_square(center = c(0,0), side = 100, col = cols[2], fill = cols_trans[2], lty = 2)
make_square(center = c(50,-50), side = 100, col = cols_trans[3], fill = cols_trans[3])
make_square(center = c(50,-50), side = 100, col = cols[1], fill = cols_trans[3])
make_square(center = c(0,0), side = 150, col = cols_trans[1], fill = cols_trans[1])
}
\author{
Stephane Guerrier
}
