% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_comparis.R
\name{get_prices}
\alias{get_prices}
\title{Get prices for a given city and page}
\usage{
get_prices(city, page, validate_page = TRUE)
}
\arguments{
\item{city}{a length one character vector indicating a city.}

\item{page}{a length one numeric vector indicating a page number.}

\item{validate_page}{a length one logical vector indicating whether or not to
validate \code{page} argument.}
}
\value{
a numeric vector of prices.
}
\description{
\code{get_prices()} scrapes prices elements for a given Swiss city and page
from \url{https://www.comparis.ch} web-page and returns a numeric vector of
prices.
}
\details{
Note that only the latest (no older than one week) items considered.
\url{https://www.comparis.ch} url's indexing starts from zero, therefore, the
page number should be subtracted by one.
}
\examples{
prices <- get_prices("bern", 3)

}
\author{
Iegor Rudnytskyi, \email{iegor.rudnytskyi@unil.ch}
}
